Compare repository visibility against the target
Now we can fetch repos, generate a target for them from the default, and
diff those. Only I think that having the "visibility" target in there is
already too verbose, we have a few dozen public repos. Maybe we should
just leave that property un-managed. But we can manage teams that have
access in the same way.