Use Nix flakes, use built-in tomllib
Flakes are said to be the future ... the convenient "nix run" command
from Nix 2.3 has been repurposed in later versions and there is no
direct replacement. 2.3 is quite old by now, so jump on the flakes
bandwagon.

This also changes the Python script to no longer use the external
"tomli" package, but to use the built-in "tomllib" instead (which is the
former "tomli" brought into the stdlib).